+++ src/main/java/lermitage/intellij/extra/icons/cfg/SettingsForm.java
@@ -2,6 +2,7 @@
 
 package lermitage.intellij.extra.icons.cfg;
 
+import com.google.common.base.Strings;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.event.DocumentEvent;
 import com.intellij.openapi.editor.event.DocumentListener;
@@ -39,6 +40,7 @@ import javax.swing.JTabbedPane;
 import javax.swing.ListSelectionModel;
 import javax.swing.RowFilter;
 import javax.swing.table.DefaultTableModel;
+import javax.swing.table.TableColumnModel;
 import javax.swing.table.TableModel;
 import javax.swing.table.TableRowSorter;
 import javax.swing.table.TableStringConverter;
@@ -85,6 +87,8 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
     private boolean isProjectForm = false;
     private List<Model> customModels = new ArrayList<>();
 
+    private boolean forceUpdate = false;
+
     public SettingsForm() {
         buttonEnableAll.addActionListener(e -> enableAll(true));
         buttonDisableAll.addActionListener(e -> enableAll(false));
@@ -129,6 +133,9 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
 
     @Override
     public boolean isModified() {
+        if (forceUpdate) {
+            return true;
+        }
         SettingsService service = SettingsService.getInstance(project);
         if (isProjectForm) {
             SettingsProjectService projectService = (SettingsProjectService) service;
@@ -203,6 +210,8 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
         } else {
             ProjectUtils.refreshOpenedProjects();
         }
+
+        forceUpdate = false;
     }
 
     @Nullable
@@ -223,8 +232,8 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
         filterTextField.setToolTipText("<html>Regex is a <b>Java regex</b> and <b>is not case-sensitive</b><br>" +
             "You can also type <b>yes</b> or <b>no</b> to find the icons that are enabled or disabled.</html>");
         filterResetBtn.setText("Reset filter");
-        bottomTip.setText("<html><b>Icons are ordered by priority</b>. To use an <i>alternative</i> icon (as for Markdown files), " +
-            "deactivate the icon(s) with higher priority.</html>");
+        bottomTip.setText("<html><b>Icons are ordered by priority</b>. To use an <b>alternative</b> icon (as for Markdown files), " +
+            "deactivate the icon(s) with higher priority.<br/>The <b>Restart</b> column indicates if you need to restart the IDE to see changes.</html>");
         initCheckbox();
         loadPluginIconsTable();
         userIconsTable.setShowHorizontalLines(false);
@@ -309,11 +318,12 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
         userIconsTable.setModel(userIconsSettingsTableModel);
         userIconsTable.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
         userIconsTable.setRowHeight(28);
-        userIconsTable.getColumnModel().getColumn(UserIconsSettingsTableModel.ICON_COL_NUMBER).setMaxWidth(28);
-        userIconsTable.getColumnModel().getColumn(UserIconsSettingsTableModel.ICON_COL_NUMBER).setWidth(28);
-        userIconsTable.getColumnModel().getColumn(UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setWidth(28);
-        userIconsTable.getColumnModel().getColumn(UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setMaxWidth(28);
-        userIconsTable.getColumnModel().getColumn(UserIconsSettingsTableModel.ICON_LABEL_COL_NUMBER).sizeWidthToFit();
+        TableColumnModel columnModel = userIconsTable.getColumnModel();
+        columnModel.getColumn(UserIconsSettingsTableModel.ICON_COL_NUMBER).setMaxWidth(28);
+        columnModel.getColumn(UserIconsSettingsTableModel.ICON_COL_NUMBER).setWidth(28);
+        columnModel.getColumn(UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setWidth(28);
+        columnModel.getColumn(UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setMaxWidth(28);
+        columnModel.getColumn(UserIconsSettingsTableModel.ICON_LABEL_COL_NUMBER).sizeWidthToFit();
         if (currentSelected != -1 && currentSelected < userIconsTable.getRowCount()) {
             userIconsTable.setRowSelectionInterval(currentSelected, currentSelected);
         }
@@ -405,6 +415,7 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
                 !disabledModelIds.contains(m.getId()),
                 m.getDescription(),
                 Arrays.toString(m.getTags().stream().map(ModelTag::getName).toArray()).replaceAll("\\[|]*", "").trim(),
+                !Strings.isNullOrEmpty(m.getIdeIcon()),
                 m.getTags(),
                 m.getId()
             })
@@ -412,16 +423,19 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
         pluginIconsTable.setModel(pluginIconsSettingsTableModel);
         pluginIconsTable.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
         pluginIconsTable.setRowHeight(28);
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_COL_NUMBER).setMaxWidth(28);
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_COL_NUMBER).setWidth(28);
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setWidth(28);
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setMaxWidth(28);
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_LABEL_COL_NUMBER).sizeWidthToFit();
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_TAGS_LABEL_COL_NUMBER).setMaxWidth(120);
-        pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_TAGS_LABEL_COL_NUMBER).setMinWidth(120);
+        TableColumnModel columnModel = pluginIconsTable.getColumnModel();
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_COL_NUMBER).setMaxWidth(28);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_COL_NUMBER).setWidth(28);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setWidth(28);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER).setMaxWidth(28);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_LABEL_COL_NUMBER).sizeWidthToFit();
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_TAGS_LABEL_COL_NUMBER).setMaxWidth(120);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_TAGS_LABEL_COL_NUMBER).setMinWidth(120);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_REQUIRE_IDE_RESTART).setMaxWidth(65);
+        columnModel.getColumn(PluginIconsSettingsTableModel.ICON_REQUIRE_IDE_RESTART).setMinWidth(65);
         // set invisible but keep data
-        pluginIconsTable.getColumnModel().removeColumn(pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_ID_COL_NUMBER));
-        pluginIconsTable.getColumnModel().removeColumn(pluginIconsTable.getColumnModel().getColumn(PluginIconsSettingsTableModel.ICON_TAGS_ENUM_LIST_COL_NUMBER));
+        columnModel.removeColumn(columnModel.getColumn(PluginIconsSettingsTableModel.ICON_ID_COL_NUMBER));
+        columnModel.removeColumn(columnModel.getColumn(PluginIconsSettingsTableModel.ICON_TAGS_ENUM_LIST_COL_NUMBER));
         if (currentSelected != -1) {
             pluginIconsTable.setRowSelectionInterval(currentSelected, currentSelected);
         }
@@ -437,6 +451,7 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
 
     private JComponent createToolbarDecorator() {
         ToolbarDecorator decorator = ToolbarDecorator.createDecorator(userIconsTable)
+
             .setAddAction(anActionButton -> {
                 ModelDialog modelDialog = new ModelDialog(this);
                 if (modelDialog.showAndGet()) {
@@ -445,7 +460,9 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
                     foldersFirst(customModels);
                     setUserIconsTableModel();
                 }
-            }).setEditAction(anActionButton -> {
+            })
+
+            .setEditAction(anActionButton -> {
                 int currentSelected = userIconsTable.getSelectedRow();
                 ModelDialog modelDialog = new ModelDialog(this);
                 modelDialog.setModelToEdit(customModels.get(currentSelected));
@@ -454,11 +471,55 @@ public class SettingsForm implements Configurable, Configurable.NoScroll {
                     customModels.set(currentSelected, newModel);
                     setUserIconsTableModel();
                 }
-            }).setRemoveAction(anActionButton -> {
+            })
+
+            .setRemoveAction(anActionButton -> {
                 customModels.remove(userIconsTable.getSelectedRow());
                 setUserIconsTableModel();
-            }).setButtonComparator("Add", "Edit", "Remove");
+            })
+
+            .setButtonComparator("Add", "Edit", "Remove")
+
+            .setMoveUpAction(anActionButton -> reorderUserIcons(MoveDirection.UP, userIconsTable.getSelectedRow()))
+
+            .setMoveDownAction(anActionButton -> reorderUserIcons(MoveDirection.DOWN, userIconsTable.getSelectedRow()));
         return decorator.createPanel();
+    }
+
+    private void reorderUserIcons(MoveDirection moveDirection, int selectedItemIdx) {
+        Model modelToMove = customModels.get(selectedItemIdx);
+        int newSelectedItemIdx = moveDirection == MoveDirection.UP ? selectedItemIdx -1 : selectedItemIdx + 1;
+        boolean selectedItemIsEnabled = (boolean) userIconsTable.getValueAt(selectedItemIdx, UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER);
+        boolean newSelectedItemIsEnabled = (boolean) userIconsTable.getValueAt(newSelectedItemIdx, UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER);
+        List<Boolean> itemsAreEnabled = new ArrayList<>();
+        for (int i = 0; i < userIconsTable.getRowCount(); i++) {
+            itemsAreEnabled.add((Boolean) userIconsTable.getValueAt(i, UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER));
+        }
+
+        customModels.set(selectedItemIdx, customModels.get(newSelectedItemIdx));
+        customModels.set(newSelectedItemIdx, modelToMove);
+        setUserIconsTableModel();
+
+        // User may have enabled or disabled some items, but changes are not applied yet to customModels, so setUserIconsTableModel will reset
+        // the Enabled column to previous state. We need to reapply user changes on this column.
+        for (int i = 0; i < userIconsTable.getRowCount(); i++) {
+            userIconsTable.setValueAt(itemsAreEnabled.get(i), i, UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER);
+        }
+        userIconsTable.setValueAt(selectedItemIsEnabled, newSelectedItemIdx, UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER);
+        userIconsTable.setValueAt(newSelectedItemIsEnabled, selectedItemIdx, UserIconsSettingsTableModel.ICON_ENABLED_COL_NUMBER);
+
+        userIconsTable.clearSelection();
+        userIconsTable.setRowSelectionInterval(newSelectedItemIdx, newSelectedItemIdx);
+
+        // TODO fix Model and ModelCondition equals & hashCode methods in order
+        //  to fix CollectionUtils.isEqualCollection(customModels, service.getCustomModels()).
+        //  For now, the comparison returns true when customModels ordering changed. It should return false.
+        forceUpdate = true;
+    }
+
+    private enum MoveDirection {
+        UP,
+        DOWN
     }
 
     private void foldersFirst(List<Model> models) {